+++ src/modules/esm.mjs
@@ -1,0 +1,36 @@
+import path from 'path';
+import { release, version } from 'os';
+import { createServer as createServerHttp } from 'http';
+import './files/c.js';
+import { readFile } from 'fs/promises';
+
+const random = Math.random();
+const __filename = import.meta.filename;
+const __dirname = import.meta.dirname;
+
+const unknownObject = await readFile(
+  path.resolve(__dirname, 'files', random > 0.5 ? 'a.json' : 'b.json'),
+  'utf-8'
+).then((data) => JSON.parse(data));
+
+console.log(`Release ${release()}`);
+console.log(`Version ${version()}`);
+console.log(`Path segment separator is "${path.sep}"`);
+
+console.log(`Path to current file is ${__filename}`);
+console.log(`Path to current directory is ${__dirname}`);
+
+const myServer = createServerHttp((_, res) => {
+  res.end('Request accepted');
+});
+
+const PORT = 3000;
+
+console.log(unknownObject);
+
+myServer.listen(PORT, () => {
+  console.log(`Server is listening on port ${PORT}`);
+  console.log('To terminate it, use Ctrl+C combination');
+});
+
+export { unknownObject, myServer };
